timing fix + idiot-proofing
New function checkDependencies() will check to see if ToyWebBridge and Intiface Central are running, and will disable most functions if they aren't.

Tweaked the timing on createPayload()'s stop sequence in order to give it some breathing room and included a new optional input that'll set the loop flag in the payload to what the user wants. Defaults to false as it's meant to be called every beatHit().

Added a cooldown to sendPayload() to prevent spamming the frontend on frame drops.

Fixed my fuckup in naming set_intensity, many apologies i am extremely mediocre.

Testing was conducted on an xbox controller and an indev build of Cracked Catastrophe.